Fix crash on unknown DDHCP command
The payload_tree is not yet instantiated when the command is looked up.
If the lookup is unsuccesful the dissector adds an error message to the
tree. However this error message is added to the payload_tree not the
main dissector tree.
Trying to reference the payload_tree at theis point causes a runtime
error.
Fix that by adding the error to the main parser tree instead.